Safari-611.1.10/LayoutTests/fast/dom/Window/postMessage-clone-port-error-expected.txt

This test checks if DataCloneError exception is thrown when postMessage's second parameter is the source port or the target port.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS dataCloneErrorWasThrown is true
PASS successfullyParsed is true

TEST COMPLETE
